+++ R/PivotDataGroup.R
@@ -36,6 +36,8 @@
 #'   associated with the PivotTable.
 #' @field style A PivotStyle object that can apply overrides to the base style
 #'   for this data group.
+#' @field isMatch Whether or not this data group matches the criteria of the
+#'   last find.
 #' @field isRendered Whether or not this data group has been rendered yet (used
 #'   as part of the rendering routines).
 #' @field isWithinVisibleRange whether or not this data group is visible (used
@@ -84,6 +86,17 @@
 #'   \item{\code{normaliseDataGroup()}}{Normalise the data group hierachy so
 #'   that all branches have the same number of levels - accomplished by adding
 #'   empty child data groups where needed.}
+#'   \item{\code{getNetFilters()}}{Get a PivotFilters object that contains all
+#'   of the filters applied in this data group and all of its ancestors.}
+#'   \item{\code{getNetCalculationName()}}{Get the calculation name applied in
+#'   this data group or its nearest ancestor.}
+#'   \item{\code{isFindMatch(variableNames=NULL, variableValues=NULL,
+#'   totals="include", calculationNames=NULL)}}{Tests whether this data group
+#'   matches the specified criteria.}
+#'   \item{\code{findDataGroups(variableNames=NULL, variableValues=NULL,
+#'   totals="include", calculationNames=NULL,
+#'   includeChildGroups=FALSE)}}{Searches all data groups underneath this data
+#'   group to find groups that match the specified criteria.}
 #'   \item{\code{asList()}}{Get a list representation of the data group(s).}
 #'   \item{\code{asJSON()}}{Get a JSON representation of the data group(s).}
 #' }
@@ -104,10 +117,10 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      checkArgument("PivotDataGroup", "initialize", calculationGroupName, missing(calculationGroupName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
      checkArgument("PivotDataGroup", "initialize", calculationName, missing(calculationName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
      private$p_parentPivot <- parentPivot
-     private$p_parentPivot$message("DataGroup$new", "Creating new data group...",
+     private$p_parentPivot$message("PivotDataGroup$new", "Creating new data group...",
                                    list(rowOrColumn=rowOrColumn, caption=caption, variableName=variableName, values=values,
                                         calculationGroupName=calculationGroupName, calculationName=calculationName))
-     if(!(rowOrColumn %in% c("row", "column"))) stop("DataGroup$new(): rowOrColumn must be either row or column", call. = FALSE)
+     if(!(rowOrColumn %in% c("row", "column"))) stop("PivotDataGroup$new(): rowOrColumn must be either row or column", call. = FALSE)
      private$p_parentGroup <- parentGroup
      private$p_rowOrColumn <- rowOrColumn
      private$p_caption <- caption
@@ -117,12 +130,12 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      private$p_groups <- list() # child groups
      private$p_calculationGroupName <- calculationGroupName
      private$p_calculationName <- calculationName
-     private$p_parentPivot$message("DataGroup$new", "Created new data group.")
+     private$p_parentPivot$message("PivotDataGroup$new", "Created new data group.")
    },
    getAncestorGroups = function(ancestors=NULL, includeCurrentGroup=FALSE) {
      checkArgument("PivotDataGroup", "getAncestorGroups", ancestors, missing(ancestors), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", allowedListElementClasses="PivotDataGroup")
      checkArgument("PivotDataGroup", "getAncestorGroups", includeCurrentGroup, missing(includeCurrentGroup), allowMissing=TRUE, allowNull=TRUE, allowedClasses="logical")
-     private$p_parentPivot$message("DataGroup$getAncestorGroups", "Getting ancestors...",
+     private$p_parentPivot$message("PivotDataGroup$getAncestorGroups", "Getting ancestors...",
                                    list(ancestorCount=length(ancestors), includeCurrentGroup=includeCurrentGroup))
      acs <- NULL
      if(missing(ancestors)||is.null(ancestors)) {
@@ -135,13 +148,13 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        acs[[index]] <- private$p_parentGroup
        acs <- private$p_parentGroup$getAncestorGroups(acs)
      }
-     private$p_parentPivot$message("DataGroup$getAncestorGroups", "Got ancestors.", list(count=length(acs)))
+     private$p_parentPivot$message("PivotDataGroup$getAncestorGroups", "Got ancestors.", list(count=length(acs)))
      return(invisible(acs)) # note the top-most parent will be at the bottom of the return list
    },
    getDescendantGroups = function(descendants=NULL, includeCurrentGroup=FALSE) {
      checkArgument("PivotDataGroup", "getDescendantGroups", descendants, missing(descendants), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", allowedListElementClasses="PivotDataGroup")
      checkArgument("PivotDataGroup", "getDescendantGroups", includeCurrentGroup, missing(includeCurrentGroup), allowMissing=TRUE, allowNull=TRUE, allowedClasses="logical")
-     private$p_parentPivot$message("DataGroup$getDescendantGroups", "Getting descendant groups...")
+     private$p_parentPivot$message("PivotDataGroup$getDescendantGroups", "Getting descendant groups...")
      dgs <- NULL
      if(missing(descendants)||is.null(descendants)) { dgs <- list() }
      else { dgs <- descendants }
@@ -154,12 +167,12 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          dgs <- private$p_groups[[i]]$getDescendantGroups(dgs, includeCurrentGroup=TRUE)
        }
      }
-     private$p_parentPivot$message("DataGroup$getDescendantGroups", "Got descendant groups", list(count=length(dgs)))
+     private$p_parentPivot$message("PivotDataGroup$getDescendantGroups", "Got descendant groups", list(count=length(dgs)))
      return(invisible(dgs))
    },
    getLeafGroups = function(leafGroups=NULL) {
      checkArgument("PivotDataGroup", "getLeafGroups", leafGroups, missing(leafGroups), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", allowedListElementClasses="PivotDataGroup")
-     private$p_parentPivot$message("DataGroup$getLeafGroups", "Getting leaf groups...", list(leafGroupCount=length(leafGroups)))
+     private$p_parentPivot$message("PivotDataGroup$getLeafGroups", "Getting leaf groups...", list(leafGroupCount=length(leafGroups)))
      lgs <- NULL
      if(missing(leafGroups)||is.null(leafGroups)) { lgs <- list() }
      else { lgs <- leafGroups }
@@ -172,13 +185,13 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        index <- length(lgs) + 1
        lgs[[index]] <- self
      }
-     private$p_parentPivot$message("DataGroup$getLeafGroups", "Got leaf groups", list(count=length(lgs)))
+     private$p_parentPivot$message("PivotDataGroup$getLeafGroups", "Got leaf groups", list(count=length(lgs)))
      return(invisible(lgs))
    },
    getLevelGroups = function(level=NULL, levelGroups=NULL) { #level=0 is the current data group
      checkArgument("PivotDataGroup", "getLevelGroups", level, missing(level), allowMissing=FALSE, allowNull=FALSE, allowedClasses=c("integer", "numeric"))
      checkArgument("PivotDataGroup", "getLevelGroups", levelGroups, missing(levelGroups), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", allowedListElementClasses="PivotDataGroup")
-     private$p_parentPivot$message("DataGroup$getLevelGroups", "Getting level groups...",
+     private$p_parentPivot$message("PivotDataGroup$getLevelGroups", "Getting level groups...",
                                    list(level=level, levelGroupCount=length(levelGroups)))
      lgs <- NULL
      if(missing(levelGroups)||is.null(levelGroups)) { lgs <- list() }
@@ -192,7 +205,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          lgs <- private$p_groups[[i]]$getLevelGroups(level-1, lgs)
        }
      }
-     private$p_parentPivot$message("DataGroup$getLevelGroups", "Got level groups", list(count=length(lgs)))
+     private$p_parentPivot$message("PivotDataGroup$getLevelGroups", "Got level groups", list(count=length(lgs)))
      return(invisible(lgs))
    },
    addChildGroup = function(variableName=NULL, values=NULL,
@@ -204,7 +217,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      checkArgument("PivotDataGroup", "addChildGroup", isTotal, missing(isTotal), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
      checkArgument("PivotDataGroup", "addChildGroup", calculationGroupName, missing(calculationGroupName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
      checkArgument("PivotDataGroup", "addChildGroup", calculationName, missing(calculationName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
-     private$p_parentPivot$message("DataGroup$addChildGroup", "Adding child group...",
+     private$p_parentPivot$message("PivotDataGroup$addChildGroup", "Adding child group...",
                                    list(caption=caption, isTotal=isTotal, variableName=variableName, values=values,
                                    calculationGroupName=calculationGroupName, calculationName=calculationName))
      private$p_parentPivot$resetCells()
@@ -215,7 +228,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
                           calculationGroupName=calculationGroupName, calculationName=calculationName)
      index <- length(private$p_groups) + 1
      private$p_groups[[index]] <- grp
-     private$p_parentPivot$message("DataGroup$addChildGroup", "Added child group.")
+     private$p_parentPivot$message("PivotDataGroup$addChildGroup", "Added child group.")
      return(invisible(grp))
    },
    # permutations:
@@ -243,7 +256,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      checkArgument("DataGroup", "addDataGroups", visualTotals, missing(visualTotals), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
      checkArgument("DataGroup", "addDataGroups", totalPosition, missing(totalPosition), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character", allowedValues=c("before", "after"))
      checkArgument("DataGroup", "addDataGroups", totalCaption, missing(totalCaption), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character")
-     private$p_parentPivot$message("DataGroup$addDataGroups", "Adding data groups...",
+     private$p_parentPivot$message("PivotDataGroup$addDataGroups", "Adding data groups...",
                                    list(variableName=variableName, atLevel=atLevel, fromData=fromData,
                                         dataName=dataName, dataSortOrder=dataSortOrder, dataFormat=dataFormat,
                                         onlyCombinationsThatExist=onlyCombinationsThatExist,
@@ -251,7 +264,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
                                         expandExistingTotals=expandExistingTotals, addTotal=addTotal, visualTotals=visualTotals,
                                         totalPosition=totalPosition, totalCaption=totalCaption))
      private$p_parentPivot$resetCells()
-     if(missing(variableName)||is.null(variableName)) stop("DataGroup$addDataGroups(): variableName must be specified", call. = FALSE)
+     if(missing(variableName)||is.null(variableName)) stop("PivotDataGroup$addDataGroups(): variableName must be specified", call. = FALSE)
      if(addTotal==TRUE){ private$p_visualTotals <- visualTotals }
      df <- NULL
      topLevelDisinctValues <- NULL
@@ -260,17 +273,17 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      if(fromData==TRUE) {
        # check that a data frame has been specified (or that we have a default data frame)
        if(missing(dataName)||is.null(dataName)) {
-         if (private$p_parentPivot$data$count < 1) stop("DataGroup$addDataGroups():  No data frames.  Specify data before calling addLeafGroup.", call. = FALSE)
+         if (private$p_parentPivot$data$count < 1) stop("PivotDataGroup$addDataGroups():  No data frames.  Specify data before calling addLeafGroup.", call. = FALSE)
          df <- private$p_parentPivot$data$defaultData
        }
        else {
          df <- private$p_parentPivot$data$getData(dataName)
-         if(is.null(df)) stop(paste0("DataGroup$addDataGroups():  No data frame found in PivotTable with name '", dataName, "'."), call. = FALSE)
+         if(is.null(df)) stop(paste0("PivotDataGroup$addDataGroups():  No data frame found in PivotTable with name '", dataName, "'."), call. = FALSE)
        }
      }
      else {
-       if (missing(explicitListOfValues)) stop("DataGroup$addDataGroups():  An explicitListOfValues must be specified when fromData=FALSE", call. = FALSE)
-       if (is.null(explicitListOfValues)) stop("DataGroup$addDataGroups():  explicitListOfValues must not be null when fromData=FALSE", call. = FALSE)
+       if (missing(explicitListOfValues)) stop("PivotDataGroup$addDataGroups():  An explicitListOfValues must be specified when fromData=FALSE", call. = FALSE)
+       if (is.null(explicitListOfValues)) stop("PivotDataGroup$addDataGroups():  explicitListOfValues must not be null when fromData=FALSE", call. = FALSE)
        topLevelDisinctValues <- explicitListOfValues
        topLevelCaptions <- names(explicitListOfValues)
        fvals <- topLevelDisinctValues
@@ -349,6 +362,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          if (rowColFilters$count > 0)
          {
            filterCmd <- NULL
+           filterCount <- 0
            for(j in 1:length(rowColFilters$filters)) {
              filter <- rowColFilters$filters[[j]]
              if(is.null(filter$variableName)) stop("PivotCalculator$getFilteredDataFrame(): filter$variableName must not be null", call. = FALSE)
@@ -358,10 +372,13 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
              if(length(filter$values)>0) {
                # %in% handles NA correctly for our use-case, i.e. NA %in% NA returns TRUE, not NA
                filterCmd <- paste0(filterCmd, "(", filter$variableName, " %in% rowColFilters$filters[[", j, "]]$values)")
+               filterCount <- filterCount + 1
              }
            }
-           filterCmd <- paste0("data <- dplyr::filter(data,", filterCmd, ")")
-           eval(parse(text=filterCmd))
+           if(filterCount > 0) {
+             filterCmd <- paste0("data <- dplyr::filter(data,", filterCmd, ")")
+             eval(parse(text=filterCmd))
+           }
          }
          # get the distinct values for the current variable
          eval(parse(text=paste0("data <- dplyr::select(data, ", variableName, ")")))
@@ -390,7 +407,8 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        # append the child groups
        else if("list" %in% class(distinctValues)) {
          if((addTotal==TRUE)&&(totalPosition=="before")) {
-           newGrp <- grp$addChildGroup(caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
+           newGrp <- grp$addChildGroup(variableName=variableName, values=NULL, # so that the totals have a reference to the variable
+                                       caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
            index <- length(newGroups) + 1
            newGroups[[index]] <- newGrp
          }
@@ -404,14 +422,16 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
            newGroups[[index]] <- newGrp
          }
          if((addTotal==TRUE)&&(totalPosition=="after")) {
-           newGrp <- grp$addChildGroup(caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
+           newGrp <- grp$addChildGroup(variableName=variableName, values=NULL, # so that the totals have a reference to the variable
+                                       caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
            index <- length(newGroups) + 1
            newGroups[[index]] <- newGrp
          }
        }
        else {
          if((addTotal==TRUE)&&(totalPosition=="before")) {
-           newGrp <- grp$addChildGroup(caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
+           newGrp <- grp$addChildGroup(variableName=variableName, values=NULL, # so that the totals have a reference to the variable
+                                       caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
            index <- length(newGroups) + 1
            newGroups[[index]] <- newGrp
          }
@@ -425,7 +445,8 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
            newGroups[[index]] <- newGrp
          }
          if((addTotal==TRUE)&&(totalPosition=="after")) {
-           newGrp <- grp$addChildGroup(caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
+           newGrp <- grp$addChildGroup(variableName=variableName, values=NULL, # so that the totals have a reference to the variable
+                                       caption=totalCaption, calculationGroupName=calculationGroupName, isTotal=TRUE)
            index <- length(newGroups) + 1
            newGroups[[index]] <- newGrp
          }
@@ -440,7 +461,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          if(descdnt$isTotal) descdnt$filters$setFilter(topLevelFilter, action="union")
        }
      }
-     private$p_parentPivot$message("DataGroup$addDataGroups", "Added groups.", list(count=length(newGroups)))
+     private$p_parentPivot$message("PivotDataGroup$addDataGroups", "Added groups.", list(count=length(newGroups)))
      return(invisible(newGroups))
    },
    sortDataGroups = function(levelNumber=1, orderBy="calculation", sortOrder="desc", calculationGroupName="default", calculationName=NULL) {
@@ -449,7 +470,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      checkArgument("DataGroup", "sortDataGroups", sortOrder, missing(sortOrder), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character", allowedValues=c("asc","desc"))
      checkArgument("DataGroup", "sortDataGroups", calculationGroupName, missing(calculationGroupName), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character")
      checkArgument("DataGroup", "sortDataGroups", calculationName, missing(calculationName), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
-     private$p_parentPivot$message("DataGroup$sortDataGroups", "Sorting data groups...",
+     private$p_parentPivot$message("PivotDataGroup$sortDataGroups", "Sorting data groups...",
                                    list(levelNumber=levelNumber, orderBy=orderBy, sortOrder=sortOrder,
                                         calculationGroupName=calculationGroupName, calculationName=calculationName))
      private$p_parentPivot$resetCells()
@@ -459,15 +480,15 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      calculation <- NULL
      if(orderBy=="calculation") {
        if(!private$p_parentPivot$calculationGroups$isExistingCalculationGroup(calculationGroupName))
-         stop(paste0("DataGroup$sortDataGroups():  There is no Calculation Group named '", calculationGroupName, "'"), call. = FALSE)
+         stop(paste0("PivotDataGroup$sortDataGroups():  There is no Calculation Group named '", calculationGroupName, "'"), call. = FALSE)
        calculationGroup <- private$p_parentPivot$calculationGroups$getCalculationGroup(calculationGroupName)
        if(is.null(calculationName)) {
          calculationName <- calculationGroup$defaultCalculationName
          if(is.null(calculationName))
-           stop(paste0("DataGroup$sortDataGroups():  No calculation has been specified and there is no default calculation."), call. = FALSE)
+           stop(paste0("PivotDataGroup$sortDataGroups():  No calculation has been specified and there is no default calculation."), call. = FALSE)
        }
        if(!calculationGroup$isExistingCalculation(calculationName))
-         stop(paste0("DataGroup$sortDataGroups():  There is no Calculation named '", calculationName , "' in group '", calculationGroupName, "'"), call. = FALSE)
+         stop(paste0("PivotDataGroup$sortDataGroups():  There is no Calculation named '", calculationName , "' in group '", calculationGroupName, "'"), call. = FALSE)
        calculation <- calculationGroup$getCalculation(calculationName)
      }
      # sort at this level, or a level below?
@@ -563,23 +584,23 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
                                               calculationGroupName=calculationGroupName, calculationName=calculationName)
        }
      }
-     private$p_parentPivot$message("DataGroup$sortDataGroups", "Sorted data groups.")
+     private$p_parentPivot$message("PivotDataGroup$sortDataGroups", "Sorted data groups.")
      return(invisible())
    },
    addCalculationGroups = function(calculationGroupName=NULL, atLevel=NULL) {
      checkArgument("PivotDataGroup", "addCalculationGroups", calculationGroupName, missing(calculationGroupName), allowMissing=FALSE, allowNull=FALSE, allowedClasses="character")
      checkArgument("PivotDataGroup", "addCalculationGroups", atLevel, missing(atLevel), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("integer", "numeric"))
-     private$p_parentPivot$message("DataGroup$addCalculationGroups", "Adding calculation groups...")
+     private$p_parentPivot$message("PivotDataGroup$addCalculationGroups", "Adding calculation groups...")
      private$p_parentPivot$resetCells()
      private$p_parentPivot$calculationsPosition <- private$p_rowOrColumn # will throw an error if trying to add calcs to one axis when already present on the other axis
-     if(missing(calculationGroupName)) stop("DataGroup$addCalculationGroups():  calculationGroupName must be specified.", call. = FALSE)
-     if(is.null(calculationGroupName)) stop("DataGroup$addCalculationGroups():  calculationGroupName cannot be null.", call. = FALSE)
+     if(missing(calculationGroupName)) stop("PivotDataGroup$addCalculationGroups():  calculationGroupName must be specified.", call. = FALSE)
+     if(is.null(calculationGroupName)) stop("PivotDataGroup$addCalculationGroups():  calculationGroupName cannot be null.", call. = FALSE)
      if(!private$p_parentPivot$calculationGroups$isExistingCalculationGroup(calculationGroupName))
-       stop(paste0("DataGroup$addCalculationGroups():  There is no Calculation Group named '", calculationGroupName, "' in the Pivot Table."), call. = FALSE)
+       stop(paste0("PivotDataGroup$addCalculationGroups():  There is no Calculation Group named '", calculationGroupName, "' in the Pivot Table."), call. = FALSE)
      # get the calculation group and the calculations to be displayed
      calculationGroup <- private$p_parentPivot$calculationGroups$getCalculationGroup(calculationGroupName)
      if(calculationGroup$count==0)
-       stop(paste0("DataGroup$addCalculationGroups():  There are no calculations in the calculation group '", calculationGroupName, "'"), call. = FALSE)
+       stop(paste0("PivotDataGroup$addCalculationGroups():  There are no calculations in the calculation group '", calculationGroupName, "'"), call. = FALSE)
      calculations <- list()
      for (i in 1:calculationGroup$count) {
        calc <- calculationGroup$calculations[[i]]
@@ -589,7 +610,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        }
      }
      if(length(calculations)==0)
-       stop(paste0("DataGroup$addCalculationGroups():  There are no visible calculations in the calculation group '", calculationGroupName, "'"), call. = FALSE)
+       stop(paste0("PivotDataGroup$addCalculationGroups():  There are no visible calculations in the calculation group '", calculationGroupName, "'"), call. = FALSE)
      calculations <- calculations[order(names(calculations))]
      # where are the new groups being added?
      if(is.null(atLevel)) {
@@ -632,12 +653,12 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          }
        }
      }
-     private$p_parentPivot$message("DataGroup$addCalculationGroups", "Added calculation groups.", list(count=length(newGroups)))
+     private$p_parentPivot$message("PivotDataGroup$addCalculationGroups", "Added calculation groups.", list(count=length(newGroups)))
      return(invisible(newGroups))
    },
    getLevelCount = function(includeCurrentLevel=FALSE) {
      checkArgument("PivotDataGroup", "getLevelCount", includeCurrentLevel, missing(includeCurrentLevel), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
-     private$p_parentPivot$message("DataGroup$levelCount", "Counting levels...")
+     private$p_parentPivot$message("PivotDataGroup$levelCount", "Counting levels...")
      # get the leaf level groups
      leafGroups <- self$getLeafGroups()
      # get the maximum number of parents of each group
@@ -647,11 +668,11 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        ancestors <- leafGroups[[i]]$getAncestorGroups(includeCurrentGroup=includeCurrentLevel)
        maxParents <- max(maxParents, length(ancestors))
      }
-     private$p_parentPivot$message("DataGroup$levelCount", "Counted levels.")
+     private$p_parentPivot$message("PivotDataGroup$levelCount", "Counted levels.")
      return(invisible(maxParents))
    },
    normaliseDataGroup = function() {
-     private$p_parentPivot$message("DataGroup$normaliseDataGroup", "Normalising data group...")
+     private$p_parentPivot$message("PivotDataGroup$normaliseDataGroup", "Normalising data group...")
      private$p_parentPivot$resetCells()
      # get the leaf level groups
      leafGroups <- self$getLeafGroups()
@@ -675,8 +696,115 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
          }
        }
      }
-     private$p_parentPivot$message("DataGroup$normaliseDataGroup", "Normalised data group.")
+     private$p_parentPivot$message("PivotDataGroup$normaliseDataGroup", "Normalised data group.")
      return(invisible())
+   },
+   getNetFilters = function() { # start at the root and apply the filter criteria that gets set at each lower level
+     private$p_parentPivot$message("PivotDataGroup$getNetFilters", "Getting net filters...")
+     ancestors <- self$getAncestorGroups(includeCurrentGroup=TRUE)
+     netFilters <- PivotFilters$new(private$p_parentPivot)
+     for(j in length(ancestors):1) {
+       acs <- ancestors[[j]]
+       filters <- acs$filters
+       if(is.null(filters)) next
+       if(filters$count==0) next
+       for(k in 1:length(filters$filters)) {
+         filter <- filters$filters[[k]]
+         netFilters$setFilter(filter, action="replace")
+       }
+     }
+     private$p_parentPivot$message("PivotDataGroup$findDataGroups", "Got net filters.")
+     return(invisible(netFilters))
+   },
+   getNetCalculationName = function() { # start at the current node and work upwards until the first calculation is encountered
+     private$p_parentPivot$message("PivotDataGroup$getNetCalculationName", "Getting net calculation...")
+     grp <- self
+     while (!is.null(grp)) {
+       if(!is.null(grp$calculationName)) return(invisible(grp$calculationName))
+       grp <- grp$parentGroup
+     }
+     private$p_parentPivot$message("PivotDataGroup$getNetCalculationName", "Got net calculation.")
+     return(invisible())
+   },
+   isFindMatch = function(variableNames=NULL, variableValues=NULL, totals="include", calculationNames=NULL) {
+     checkArgument("PivotDataGroup", "isFindMatch", variableNames, missing(variableNames), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
+     checkArgument("PivotDataGroup", "isFindMatch", variableValues, missing(variableValues), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", listElementsMustBeAtomic=TRUE)
+     checkArgument("PivotDataGroup", "isFindMatch", totals, missing(totals), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character", allowedValues=c("include", "exclude", "only"))
+     checkArgument("PivotDataGroup", "isFindMatch", calculationNames, missing(calculationNames), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
+     private$p_parentPivot$message("PivotDataGroup$isFindMatch", "Checking if matches criteria...")
+     # a) check the filter match
+     if((!is.null(variableNames))||(!is.null(variableValues))) {
+       netFilters <- self$getNetFilters()
+       if(is.null(netFilters)) return(invisible(FALSE))
+       isMatch <- netFilters$isFilterMatch(variableNames=variableNames, variableValues=variableValues)
+       if(isMatch==FALSE) return(invisible(FALSE))
+     }
+     # b) check totals criteria
+     if((totals=="exclude")&&(self$isTotal==TRUE)) return(invisible(FALSE))
+     if((totals=="only")&&(self$isTotal==FALSE)) return(invisible(FALSE))
+     # c) check calculation criteria
+     if(!is.null(calculationNames)) {
+       calcName <- self$getNetCalculationName()
+       if(is.null(calcName)) return(invisible(FALSE))
+       if(!(calcName %in% calculationNames)) return(invisible(FALSE))
+     }
+     private$p_parentPivot$message("PivotDataGroup$isFindMatch", "Checked if matches criteria.")
+     return(invisible(TRUE))
+   },
+   findDataGroups = function(variableNames=NULL, variableValues=NULL, totals="include",
+                             calculationNames=NULL, includeChildGroups=FALSE) {
+     checkArgument("PivotDataGroup", "findDataGroups", variableNames, missing(variableNames), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
+     checkArgument("PivotDataGroup", "findDataGroups", variableValues, missing(variableValues), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", listElementsMustBeAtomic=TRUE)
+     checkArgument("PivotDataGroup", "findDataGroups", totals, missing(totals), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character", allowedValues=c("include", "exclude", "only"))
+     checkArgument("PivotDataGroup", "findDataGroups", calculationNames, missing(calculationNames), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
+     checkArgument("PivotDataGroup", "findDataGroups", includeChildGroups, missing(includeChildGroups), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
+     private$p_parentPivot$message("PivotDataGroup$findDataGroups", "Finding data groups...")
+     # clear the isMatch flag across all descendants
+     clearFlags <- function(dg) {
+       dg$isMatch <- FALSE
+     }
+     grps <- self$getDescendantGroups(includeCurrentGroup=TRUE)
+     lapply(grps, clearFlags)
+     # iterate the leaf groups...
+     leafgrps <- self$getLeafGroups()
+     if(is.null(leafgrps)) return(invisible(NULL))
+     if(length(leafgrps)==0) return(invisible(NULL))
+     for(i in 1:length(leafgrps)) {
+       # for each leaf group, start at the leaf level and work up the ancestors towards the root, checking
+       # at each level whether the node is a match.
+       matches <- list()
+       grp <- leafgrps[[i]]
+       while(!is.null(grp)) {
+         if(grp$isFindMatch(variableNames=variableNames, variableValues=variableValues, totals=totals,
+                            calculationNames=calculationNames)) {
+           matches[[length(matches)+1]] <- grp
+           grp <- grp$parentGroup
+         }
+         else break
+       }
+       # if we have a match, then set the isMatched flag
+       if(includeChildGroups==TRUE) {
+         # mark the top-most group and all groups beneath (in this branch) as a match
+         if(length(matches)>0) {
+           for(j in 1:length(matches)) {
+             matches[[j]]$isMatch <- TRUE
+           }
+         }
+       }
+       else {
+         # only mark the topmost group as a match
+         if(length(matches)>0) matches[[length(matches)]]$isMatch <- TRUE
+       }
+     }
+     # collect the matching groups
+     grps <- self$getDescendantGroups(includeCurrentGroup=TRUE)
+     matches <- list()
+     for(i in 1:length(grps)) {
+       grp <- grps[[i]]
+       if(grp$isMatch) matches[[length(matches)+1]] <- grp
+     }
+     private$p_parentPivot$message("PivotDataGroup$findDataGroups", "Found data groups.")
+     return(invisible(matches))
    },
    asList = function() {
      grps <- list()
@@ -694,7 +822,6 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        calculationGroupName = private$p_calculationGroupName,
        calculationName = private$p_calculationName,
        rowColumnNumber = private$p_rowColumnNumber,
-       isRendered = private$p_isRendered,
        baseStyleName = private$p_baseStyleName,
        style = ifelse(is.null(private$p_style), "", private$p_style$asList()),
        groups = grps
@@ -714,7 +841,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      else {
        checkArgument("PivotDataGroup", "calculationGroupName", value, missing(value), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
        if(!(private$p_parentPivot$calculationGroups$isExistingCalculationGroup(value))) {
-         stop(paste0("DataGroup$calculationGroupName(): The Calculation Group '", value, "' does not exist in the Pivot Table."), call. = FALSE)
+         stop(paste0("PivotDataGroup$calculationGroupName(): The Calculation Group '", value, "' does not exist in the Pivot Table."), call. = FALSE)
        }
        private$p_calculationGroupName <- value
        return(invisible())
@@ -725,10 +852,10 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      else {
        checkArgument("PivotDataGroup", "calculationName", value, missing(value), allowMissing=TRUE, allowNull=TRUE, allowedClasses="character")
        if(is.null(private$p_calculationGroupName))
-         stop("DataGroup$calculationName(): Specify the Calculation Group before the Calculation.", call. = FALSE)
+         stop("PivotDataGroup$calculationName(): Specify the Calculation Group before the Calculation.", call. = FALSE)
        calculationGroup <- private$p_parentPivot$calculationGroups$getCalculationGroup(private$p_calculationGroupName)
        if(!(calculationGroup$isExistingCalculation(value))) {
-         stop(paste0("DataGroup$calculationName(): The Calculation '", value,
+         stop(paste0("PivotDataGroup$calculationName(): The Calculation '", value,
                      "' does not exist in the Calculation Group '", private$p_calculationGroupName, "'"), call. = FALSE)
        }
        private$p_calculationName <- value
@@ -749,7 +876,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      else {
        checkArgument("PivotDataGroup", "rowColumnNumber", value, missing(value), allowMissing=TRUE, allowNull=TRUE, allowedClasses=c("integer", "number"))
        if(!is.null(value)) {
-         if(!is.integer(value)) stop("DataGroup$rowColumnNumber(): rowColumnNumber must be an integer", call. = FALSE)
+         if(!is.integer(value)) stop("PivotDataGroup$rowColumnNumber(): rowColumnNumber must be an integer", call. = FALSE)
        }
        private$p_rowColumnNumber <- value
        return(invisible())
@@ -768,6 +895,14 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      else {
        checkArgument("PivotDataGroup", "style", value, missing(value), allowMissing=FALSE, allowNull=FALSE, allowedClasses="PivotStyle")
        private$p_style <- value
+       return(invisible())
+     }
+   },
+   isMatch = function(value) {
+     if(missing(value)) { return(invisible(private$p_isMatch)) }
+     else {
+       checkArgument("PivotDataGroup", "isMatch", value, missing(value), allowMissing=TRUE, allowNull=TRUE, allowedClasses="logical")
+       private$p_isMatch <- value
        return(invisible())
      }
    },
@@ -832,6 +967,7 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
    p_rowColumnNumber = NULL,
    p_baseStyleName = NULL,
    p_style = NULL,
+   p_isMatch = FALSE, # helper flag used when searching through data groups
    p_isWithinVisibleRange = TRUE, # helper flag to allow a subsection of the pivot table to be output
    p_isRendered = FALSE, # helper flag to keep track of which data groups have already been rendered